Zajęcia 8 + 10

+++ Zajęcia 8/1.c
@@ -2,9 +2,9 @@
 
 struct wektor3D
 {
-  int x;
-  int y;
-  int z;
+  float x;
+  float y;
+  float z;
 };
 
 struct wektor3D sumujWektor(struct wektor3D w1, struct wektor3D w2)
@@ -19,9 +19,9 @@ struct wektor3D sumujWektor(struct wektor3D w1, struct wektor3D w2)
 
 void iloczynSkalarny(struct wektor3D *w1, struct wektor3D *w2)
 {
-  int skalar = w1->x * w2->x + w1->y * w2->y + w1->z * w2->z;
+  float skalar = w1->x * w2->x + w1->y * w2->y + w1->z * w2->z;
 
-  printf("Iloczyn skalarny = %d\n", skalar);
+  printf("Iloczyn skalarny = %f\n", skalar);
 }
 
 void iloczynWektorowy(struct wektor3D w1, struct wektor3D w2, struct wektor3D *wyn)
@@ -36,18 +36,18 @@ int main(int argc, char const *argv[])
   struct wektor3D w1, w2, wynSum, wynWekt;
 
   printf("Podaj wspolrzedne wektora 1:\n");
-  scanf("%d %d %d", &(w1.x), &(w1.y), &(w1.z));
+  scanf("%f %f %f", &(w1.x), &(w1.y), &(w1.z));
 
   printf("Podaj wspolrzedne wektora 2:\n");
-  scanf("%d %d %d", &(w2.x), &(w2.y), &(w2.z));
+  scanf("%f %f %f", &(w2.x), &(w2.y), &(w2.z));
 
   wynSum = sumujWektor(w1, w2);
-  printf("Sum: x = %d, y = %d, z = %d\n", wynSum.x, wynSum.y, wynSum.z);
+  printf("Sum: x = %f, y = %f, z = %f\n", wynSum.x, wynSum.y, wynSum.z);
 
   iloczynSkalarny(&w1, &w2);
 
   iloczynWektorowy(w1, w2, &wynWekt);
-  printf("Wekt: x = %d, y = %d, z = %d\n", wynWekt.x, wynWekt.y, wynWekt.z);
+  printf("Wekt: x = %f, y = %f, z = %f\n", wynWekt.x, wynWekt.y, wynWekt.z);
 
   return 0;
 }
